+++ include/gb.h
@@ -1,31 +1,16 @@
 #ifndef __GB__
 #define __GB__
-
-#ifndef _WIN32
-#include <linux/limits.h>
-#else
-#define SDL_MAIN_HANDLED
-#endif
-
-#include <stdlib.h>
-#include <stdio.h>
-#include <string.h>
 #include <stdbool.h>
-
-#include <SDL2/SDL.h>
 
 #include "timer.h"
 #include "registers.h"
-#include "io.h"
 #include "gpu.h"
 #include "interrupt.h"
-#include "joypad.h"
 #include "memory.h"
+#include "joypad.h"
 #include "cpu.h"
 #include "joystick_config.h"
 #include "config.h"
-#include "utils.h"
-#include "log.h"
 
 #define GB_W 160
 #define GB_H 144
