+++ include/special_registers.h
@@ -4,6 +4,120 @@
 
 #ifndef INCLUDE_SPECIAL_REGISTERS_H_
 #define INCLUDE_SPECIAL_REGISTERS_H_
+#include <inttypes.h>
+#include <stdbool.h>
+
+enum sprite_size {
+	SPRITE_SIZE_8X8,
+	SPRITE_SIZE_8X16,
+};
+
+enum tile_map_display_select {
+	TILE_MAP_DISPLAY_SELECT_8,
+	TILE_MAP_DISPLAY_SELECT_C,
+};
+
+enum bg_window_tile_data_select {
+	BG_WINDOW_TILE_DATA_SELECT_8,
+	BG_WINDOW_TILE_DATA_SELECT_0,
+};
+
+static inline uint16_t tile_map_display_select_to_addr(
+		enum tile_map_display_select sel)
+{
+	static const uint16_t addr[] = {
+			[TILE_MAP_DISPLAY_SELECT_8] = 0x9800,
+			[TILE_MAP_DISPLAY_SELECT_C] = 0x9c00,
+	};
+
+	return addr[sel];
+}
+
+static inline uint16_t bg_window_tile_data_select_to_addr(
+		enum bg_window_tile_data_select sel)
+{
+	static const uint16_t addr[] = {
+			[BG_WINDOW_TILE_DATA_SELECT_8] = 0x8800,
+			[BG_WINDOW_TILE_DATA_SELECT_0] = 0x8000,
+	};
+
+	return addr[sel];
+}
+
+#pragma pack(push, 1)
+struct lcdc {
+	bool enable_bg_window_display:1;
+	bool enable_sprite_display:1;
+	enum sprite_size sprite_size:1;
+	enum tile_map_display_select bg_tile_map_display_select:1;
+	enum bg_window_tile_data_select bg_window_tile_data_select:1;
+	bool enable_window_display:1;
+	enum tile_map_display_select window_tile_map_display_select:1;
+	bool enable_lcd:1;
+} __attribute__((__packed__));
+#pragma pack(pop)
+
+#pragma pack(push, 1)
+struct spec_reg {
+	/* joypad state */
+	uint8_t p1;
+	/* serial transfer data */
+	uint8_t sb;
+	/* serial io control */
+	uint8_t sc;
+	uint8_t padding_ff03;
+	/* divide register */
+	uint8_t div;
+	/* timer counter */
+	uint8_t tima;
+	/* timer modulo */
+	uint8_t tma;
+	/* timer control */
+	uint8_t tac;
+	uint8_t padding_ff08_0e[7];
+	uint8_t ifl; /*  0xff0fu */
+	uint8_t nr10; /*  0xff10u */
+	uint8_t nr11; /*  0xff11u */
+	uint8_t nr12; /*  0xff12u */
+	uint8_t nr13; /*  0xff13u */
+	uint8_t nr14; /*  0xff14u */
+	uint8_t padding_ff15;
+	uint8_t nr21; /*  0xff16u */
+	uint8_t nr22; /*  0xff17u */
+	uint8_t nr23; /*  0xff18u */
+	uint8_t nr24; /*  0xff19u */
+	uint8_t nr30; /*  0xff1au */
+	uint8_t nr31; /*  0xff1bu */
+	uint8_t nr32; /*  0xff1cu */
+	uint8_t nr33; /*  0xff1du */
+	uint8_t nr34; /*  0xff1eu */
+	uint8_t padding_ff1f;
+	uint8_t nr41; /*  0xff20u */
+	uint8_t nr42; /*  0xff21u */
+	uint8_t nr43; /*  0xff22u */
+	uint8_t nr44; /*  0xff23u */
+	uint8_t nr50; /*  0xff24u */
+	uint8_t nr51; /*  0xff25u */
+	uint8_t nr52; /*  0xff26u */
+	uint8_t padding_ff27_2f[9];
+	uint8_t wpram[0x10];
+	union {
+		uint8_t raw; /*  0xff40u */
+		struct lcdc lcdc;
+	};
+	uint8_t stat; /*  0xff41u */
+	uint8_t scy; /*  0xff42u */
+	uint8_t scx; /*  0xff43u */
+	uint8_t ly; /*  0xff44u */
+	uint8_t lyc; /*  0xff45u */
+	uint8_t dma; /*  0xff46u */
+	uint8_t bgp; /*  0xff47u */
+	uint8_t obp0; /*  0xff48u */
+	uint8_t obp1; /*  0xff49u */
+	uint8_t wy; /*  0xff4au */
+	uint8_t wx; /*  0xff4bu */
+} __attribute__((__packed__));
+#pragma pack(pop)
 
 enum special_register {
 	SPECIAL_REGISTER_INVALID = 0u,
